requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-mathfilters==1.0.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
zope.interface==5.1.0